refactor: move ecstore format layout owners
